# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DeFarm Engines API
  version: 1.0.0
  description: |
    DeFarm Engines API provides a comprehensive agricultural supply chain tracking system
    with tokenization, circuit management, and secure item verification.

    ## Key Features
    - **JWT Authentication** - Secure token-based authentication
    - **Circuit Management** - Permission-controlled data sharing repositories
    - **Item Tokenization** - Local ID (LID) to DFID conversion via circuits
    - **Enhanced Identifiers** - Canonical and contextual identifier support
    - **Storage Adapters** - IPFS, Stellar blockchain, and custom storage backends
    - **Webhook System** - Post-action webhooks for circuit events
    - **Audit Trail** - Complete event history and activity tracking

    ## Authentication
    All authenticated endpoints require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Or API key authentication:
    ```
    X-API-Key: dfm_<your-api-key>
    ```
  contact:
    name: DeFarm API Support
    email: api@defarm.net
  license:
    name: Proprietary
servers:
  - url: https://connect.defarm.net/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in format dfm_<32-character-string>
  schemas:
...
    AdapterConfig:
      type: object
      required:
        - adapter_type
        - requires_approval
        - auto_migrate_existing
      properties:
        adapter_type:
          type: string
          description: |
            Storage adapter type. Available adapters depend on user tier:
            - "none" - No blockchain storage (all tiers)
            - "StellarTestnetIpfs" - Stellar testnet + IPFS (Professional+)
            - "StellarMainnetIpfs" - Stellar mainnet + IPFS (Enterprise only)

            Note: API accepts both hyphenated format (stellar_testnet-ipfs) and
            CamelCase format (StellarTestnetIpfs).
          enum:
            - none
            - StellarTestnetIpfs
            - StellarMainnetIpfs
          example: "StellarTestnetIpfs"
        requires_approval:
          type: boolean
          description: Whether push operations require approval. **REQUIRED field.**
          example: false
        auto_migrate_existing:
          type: boolean
          description: Auto-migrate existing items when adapter changes. **REQUIRED field.**
          example: false
        sponsor_adapter_access:
          type: boolean
          default: false
          description: Circuit sponsors adapter access for all members (overrides tier limits)
        configured_by:
          type: string
          description: User ID who configured the adapter
          readOnly: true
          example: "hen-admin-001"
        configured_at:
          type: string
          format: date-time
          description: When the adapter was configured
          readOnly: true
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "AUTHENTICATION_FAILED"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid credentials provided"
        details:
          type: object
          additionalProperties: true
          description: Additional error context
        suggestions:
          type: array
          items:
            type: string
          description: Actionable recovery suggestions
          example:
            - "Verify your username and password"
            - "Check if your account is active"
...
paths:
...
  /circuits/{circuit_id}/adapter:
    get:
      tags:
        - Circuits
      summary: Get circuit adapter configuration
      description: Retrieve the adapter configuration for a circuit
      operationId: getCircuitAdapter
      parameters:
        - name: circuit_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Adapter configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AdapterConfig'

    put:
      tags:
        - Circuits
      summary: Update circuit adapter configuration
      description: |
        Update adapter settings for a circuit. Only circuit owner or admins can configure adapters.
        The requester must have access to the specified adapter type in their tier.
        Changes are persisted to PostgreSQL immediately.
      operationId: updateCircuitAdapter
      parameters:
        - name: circuit_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the circuit to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - adapter_type
                - sponsor_adapter_access
                - requires_approval
                - auto_migrate_existing
              properties:
                adapter_type:
                  type: string
                  enum: [none, ipfs-ipfs, stellar_testnet-ipfs, stellar_mainnet-ipfs]
                  example: "stellar_mainnet-ipfs"
                sponsor_adapter_access:
                  type: boolean
                  example: true
                  description: If true, circuit pays for adapter access for all members
                requires_approval:
                  type: boolean
                  example: false
                  description: Whether migrations require approval
                auto_migrate_existing:
                  type: boolean
                  example: false
                  description: Auto-migrate items on adapter change
            examples:
              stellar_mainnet:
                summary: Configure Stellar Mainnet adapter
                value:
                  adapter_type: "stellar_mainnet-ipfs"
                  sponsor_adapter_access: false
                  requires_approval: false
                  auto_migrate_existing: false
              ipfs_sponsored:
                summary: IPFS with sponsored access
                value:
                  adapter_type: "ipfs-ipfs"
                  sponsor_adapter_access: true
                  requires_approval: false
                  auto_migrate_existing: false
      responses:
        '200':
          description: Adapter configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  adapter_type:
                    type: string
                    example: "stellar_mainnet-ipfs"
                  sponsor_adapter_access:
                    type: boolean
                    example: false
                  requires_approval:
                    type: boolean
                    example: false
                  auto_migrate_existing:
                    type: boolean
                    example: false
                  configured_by:
                    type: string
                    example: "hen-admin-001"
                  configured_at:
                    type: string
                    format: date-time
                    example: "2025-10-21T23:33:49.631504998+00:00"
        '403':
          description: Permission denied - user doesn't have adapter access or isn't circuit owner/admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...